+++ integrations/layer/serializers/conversation.js
@@ -10,5 +10,3 @@ function ConversationsSerializer(conversation) {
 }
 
 module.exports = ConversationsSerializer;
-
-